+++ indico/modules/events/registration/controllers/management/reglists_test.py
@@ -5,14 +5,19 @@
 # modify it under the terms of the MIT License; see the
 # LICENSE file for more details.
 
+from decimal import Decimal
+
 import pytest
 from flask import request
 from werkzeug.exceptions import UnprocessableEntity
 
 from indico.modules.events.registration.controllers.management.fields import _fill_form_field_with_data
-from indico.modules.events.registration.controllers.management.reglists import RHRegistrationCreate, RHRegistrationEdit
+from indico.modules.events.registration.controllers.management.reglists import (RHRegistrationCreate,
+                                                                                RHRegistrationEdit,
+                                                                                RHRegistrationsBasePrice)
 from indico.modules.events.registration.models.form_fields import RegistrationFormField
 from indico.modules.events.registration.models.items import RegistrationFormSection
+from indico.modules.events.registration.models.registrations import RegistrationState
 from indico.modules.events.registration.util import create_registration
 
 
@@ -40,9 +45,7 @@ def test_registration_edit_override_required(
         'first_name': other_user.first_name,
         'last_name': other_user.last_name,
     }
-    reg = create_registration(
-        dummy_regform, data, invitation=None, management=True, notify_user=False
-    )
+    reg = create_registration(dummy_regform, data, invitation=None, management=True, notify_user=False)
 
     # Attempting to edit the registration without setting the required field works in edit mode
     with app_context.test_request_context(json={}):
@@ -93,3 +96,68 @@ def test_registration_create_override_required(db, dummy_regform, app_context):
         rh._process_args()
         rh._process_POST()
         assert dummy_regform.active_registration_count == 1
+
+
+@pytest.mark.parametrize(
+    ('base_price_1', 'base_price_2', 'initial_state', 'action', 'new_price', 'apply_complete',
+     'expected_price', 'expected_state'),
+    (
+        (5, 10, RegistrationState.complete, 'remove', None, None, 5, RegistrationState.complete),
+        (5, 10, RegistrationState.unpaid, 'remove', None, None, 0, RegistrationState.complete),
+        (5, 10, RegistrationState.complete, 'default', None, False, 5, RegistrationState.complete),
+        (5, 10, RegistrationState.complete, 'default', None, True, 5, RegistrationState.complete),
+        (0, 10, RegistrationState.complete, 'default', None, False, 0, RegistrationState.complete),
+        (0, 10, RegistrationState.complete, 'default', None, True, 10, RegistrationState.unpaid),
+        (5, 10, RegistrationState.unpaid, 'default', None, False, 10, RegistrationState.unpaid),
+        (5, 10, RegistrationState.unpaid, 'default', None, True, 10, RegistrationState.unpaid),
+        (0, 10, RegistrationState.unpaid, 'default', None, False, 10, RegistrationState.unpaid),
+        (0, 10, RegistrationState.unpaid, 'default', None, True, 10, RegistrationState.unpaid),
+        (5, 10, RegistrationState.complete, 'custom', 15, False, 5, RegistrationState.complete),
+        (5, 10, RegistrationState.complete, 'custom', 15, True, 5, RegistrationState.complete),
+        (0, 10, RegistrationState.complete, 'custom', 15, False, 0, RegistrationState.complete),
+        (0, 10, RegistrationState.complete, 'custom', 15, True, 15, RegistrationState.unpaid),
+        (5, 10, RegistrationState.unpaid, 'custom', 15, False, 15, RegistrationState.unpaid),
+        (5, 10, RegistrationState.unpaid, 'custom', 15, True, 15, RegistrationState.unpaid),
+    ),
+)
+@pytest.mark.usefixtures('smtp')
+def test_registration_update_base_price(dummy_regform, dummy_user, app_context, base_price_1, base_price_2,
+                                        initial_state, action, new_price, apply_complete, expected_price,
+                                        expected_state):
+    # Set the base price to the first registration
+    dummy_regform.base_price = Decimal(base_price_1)
+
+    # Register a new user
+    data = {
+        'email': dummy_user.email,
+        'first_name': dummy_user.first_name,
+        'last_name': dummy_user.last_name,
+    }
+    reg = create_registration(dummy_regform, data, invitation=None, management=True, notify_user=False)
+    dummy_regform.base_price = Decimal(base_price_2)
+    reg.state = initial_state
+
+    # Generate the request data
+    form_data = {
+        'action': action,
+        'registration_id': reg.id,
+        'submitted': '',
+        'csrf_token': '00000000-0000-0000-0000-000000000000'
+    }
+    if new_price is not None:
+        form_data['base_price'] = new_price
+    if apply_complete:
+        form_data['apply_complete'] = 'y'
+    with app_context.test_request_context(method='POST', data=form_data):
+        request.view_args = {
+            'reg_form_id': dummy_regform.id,
+            'event_id': dummy_regform.event_id,
+        }
+
+        rh = RHRegistrationsBasePrice()
+        rh._process_args()
+        response = rh._process()
+        assert response.status_code == 200
+
+    assert reg.base_price == Decimal(expected_price)
+    assert reg.state == expected_state