+++ CHANGES.rst
@@ -74,6 +74,8 @@ Improvements
 - Add week and day views in the category calendar and improve navigation controls
   (:issue:`6108, 6129, 6107`, :pr:`6110`, thanks :user:`Moliholy, unconventionaldotdev`).
 - Add the ability to clone privacy settings (:pr:`6156`, thanks :user:`SegiNyn`)
+- Add option for managers to change the registration fee of a set of registrations (:issue:`6132`,
+  :pr:`6138`)
 
 Bugfixes
 ^^^^^^^^